? and # in file names.
Please see notes in issue #38, of which this is a continuation.

In short, because of a mistake in Socket.io, files with a `?` in it won't be
editable through pencil; the OT system will actually communicate all changes to
the corresponding file name up to the `?`. We have chosen to resolve this by
distancing ourselves from Socket.io.