Maintain scroll position in live Markdown app
In order for us to execute JS in the Markdown, we must create a new
document, similar to the HTML app (bbf61dcf).

(We could also execute it similarly to the Markdown app, cf. db40419b,
but this would cause it to share the app<E2><80><99>s JS environment, which can
cause variable collision and can crash the whole page.)

As a result, we also need to reset the scroll position, so that if we
are scrolled down, we keep seeing the same stuff on the right.