Maintain scroll position in HTML app
Since bbf61dcf, the HTML app no longer maintains scroll position: if the
live view is scrolled down, and a change is made in the editor, the live
view scrolls to the top again, because of the document’s initialization.

This uses the trick also used in 09ae4cb1 in the live Markdown app.